promote cli conf to its own file - add descriptions to cli options - rename exec to all
